add CommandTests, refactoring

+++ src/main/java/command/MovieDeleteCommand.java
@@ -1,10 +1,13 @@
 package command;
 
+import dto.MovieDto;
 import entities.Movie;
 import exceptions.EntityAlreadyExistException;
 import org.apache.log4j.LogManager;
 import org.apache.log4j.Logger;
 import service.MovieService;
+import web.form.IdForm;
+import web.form.validation.IdValidator;
 import web.handler.ExceptionHandler;
 
 import javax.servlet.http.HttpServletRequest;
@@ -14,21 +17,30 @@ import java.util.List;
 public class MovieDeleteCommand extends MultipleMethodCommand {
     private static final Logger LOGGER = LogManager.getLogger(MovieCommand.class);
     private MovieService movieService;
+    private IdValidator idValidator;
 
-    public MovieDeleteCommand(MovieService movieService) {
+    public MovieDeleteCommand(MovieService movieService, IdValidator idValidator) {
+
         this.movieService = movieService;
+        this.idValidator = idValidator;
     }
 
     @Override
     public String performGet(HttpServletRequest request) {
-        List<Movie> movies = movieService.findAll();
+        List<MovieDto> movies = movieService.findAllOrderBy("id");
         request.setAttribute("movie", movies);
         return "/WEB-INF/admin/movie.jsp";
     }
 
     @Override
     public String performPost(HttpServletRequest request) {
-        int id = Integer.parseInt(request.getParameter("id"));
+        String movieId = request.getParameter("id");
+        IdForm idForm = new IdForm(movieId);
+        if (idValidator.validate(idForm)) {
+            LOGGER.error("Illegal movie id");
+            throw new IllegalArgumentException();
+        }
+        int id = Integer.parseInt(movieId);
         Movie entityById = movieService.findEntityById(id);
         try {
             movieService.delete(entityById);
@@ -37,9 +49,8 @@ public class MovieDeleteCommand extends MultipleMethodCommand {
             LOGGER.info("Movie already used");
             return handler.handling(request);
         }
-        return "redirect:admin/movie" ;
+        return "redirect:admin/movie";
     }
-
 
 
 }
